# Fix access to uninitialised memory in `RawToDigi_kernel`