@fwyzard 2018-11-29
No negative impact on runtime performance.
Before:
```
Running 4 times over 4200 events with 1 jobs, each with 8 threads, 8 streams and 1 GPUs
  1833.6 ±   2.6 ev/s (4000 events)
  1840.2 ±   2.6 ev/s (4000 events)
  1812.8 ±   2.1 ev/s (4000 events)
  1848.8 ±   2.7 ev/s (4000 events)
```

After:
```
  1851.5 ±   2.1 ev/s (4000 events)
  1850.1 ±   2.8 ev/s (4000 events)
  1874.4 ±   2.7 ev/s (4000 events)
  1873.6 ±   2.2 ev/s (4000 events)
```