@fwyzard 2018-11-29
Comparison of the summaries look good.
The `initcheck` errors are gone.